+++ tests/testthat.R
@@ -1,4 +1,4 @@
 library(testthat)
 library(patrick)
-library(pfgsea)
-test_check("pfgsea")
+library(acidgsea)
+test_check("acidgsea")