+++ tests/testthat/helper-cache.R
@@ -3,7 +3,10 @@ if (!isTRUE(goalie::hasInternet())) {
     return()
 }
 dir.create("cache", showWarnings = FALSE)
-files <- "mm_deseq.rds"
+files <- c(
+    "h.all.v7.0.symbols.gmt",
+    "mm_deseq.rds"
+)
 mapply(
     FUN = function(remoteDir, file, envir) {
         destfile <- file.path("cache", file)
@@ -16,7 +19,7 @@ mapply(
     },
     file = files,
     MoreArgs = list(
-        remoteDir = pfgseaTestsURL,
+        remoteDir = acidgseaTestsURL,
         envir = environment()
     )
 )