+++ tests/testthat/test-enrichedGeneSets.R
@@ -1,19 +1,44 @@
-context("enrichedGeneSets")
+context("enrichedGeneSets : FGSEAList")
 
-test_that("FGSEAList", {
-    object <- enrichedGeneSets(gsea, collection = "h", flatten = TRUE)
-    expect_type(object, "list")
-    expect_identical(
-        object = names(object),
-        expected = c("dmso_r1881_vs_etoh_up", "dmso_r1881_vs_etoh_down")
+args <- list(
+    object = fgsea,
+    collection = "h",
+    alpha = 0.7
+)
+
+test_that("Upregulated gene sets", {
+    object <- do.call(
+        what = enrichedGeneSets,
+        args = c(args, direction = "up")
     )
 })
 
-test_that("Disable flatten mode", {
-    object <- enrichedGeneSets(gsea, collection = "h", flatten = FALSE)
+test_that("Downregulated gene sets", {
+    object <- do.call(
+        what = enrichedGeneSets,
+        args = c(args, direction = "down")
+    )
+})
+
+test_that("Enriched in both directions", {
+    object <- do.call(
+        what = enrichedGeneSets,
+        args = c(args, direction = "both")
+    )
     expect_type(object, "list")
     expect_identical(
         object = names(object),
-        expected = "dmso_r1881_vs_etoh"
+        expected = c(
+            "condition_B_vs_A",
+            "treatment_D_vs_C"
+        )
+    )
+    expect_identical(
+        object = length(object[["condition_B_vs_A"]]),
+        expected = 9L
+    )
+    expect_identical(
+        object = length(object[["treatment_D_vs_C"]]),
+        expected = 21L
     )
 })